Refactor: avoid silly future import cycles.
