+++ libs/wire-api/wire-api.cabal
@@ -235,6 +235,7 @@ library
     Wire.API.User.IdentityProvider
     Wire.API.User.Orphans
     Wire.API.User.Password
+    Wire.API.User.Phone
     Wire.API.User.Profile
     Wire.API.User.RichInfo
     Wire.API.User.Saml
